% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnome_effect.R
\name{gnome_effect}
\alias{gnome_effect}
\title{Effect Size Function}
\usage{
gnome_effect(a, c, e, g, b, varA = 1, varC = 1, varE = 1)
}
\arguments{
\item{a}{Additive genetic path coefficient}

\item{c}{Shared environment path coefficient}

\item{e}{Non-shared environmental path coefficient}

\item{g}{Path coefficient cultural transmission}

\item{b}{Path coefficient sibling interaction}

\item{varA}{Additive genetic variance}

\item{varC}{Shared environment variance}

\item{varE}{Non-shared environmental variance}
}
\value{
Proportion of variance increase
}
\description{
Function to calculate variance increase due to AC covariance
}
\examples{
gnome_effect(a = sqrt(.4), c = sqrt(.3), e = sqrt(.3), g = 0, b = sqrt(.05))
gnome_effect(a = sqrt(.4), c = sqrt(.3), e = sqrt(.3), g = sqrt(.05), b = 0)
gnome_effect(a = sqrt(.4), c = sqrt(.3), e = sqrt(.3), g = sqrt(.05), b = sqrt(.05))
}
