% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnome_mx_simulation.R
\name{gnome_mx_simulation}
\alias{gnome_mx_simulation}
\title{OpenMx Simulation Function}
\usage{
gnome_mx_simulation(
  alpha = 0.05,
  seed = NA,
  standPGS = FALSE,
  nmz = 4000,
  ndz = 4000,
  a = default_a,
  c = default_c,
  e = default_e,
  ct = default_ct,
  si = default_si,
  x = default_x,
  nloci = 100,
  npgsloci = c(2, 5, 10, 15),
  assortm = 0
)
}
\arguments{
\item{alpha}{Alpha used to calculate power}

\item{seed}{Set a seed if desired}

\item{standPGS}{Boolean to standardize the PGS}

\item{nmz}{Sample size monozygotic twins}

\item{ndz}{Sample size dizygotic twins}

\item{a}{Additive genetic path coefficient}

\item{c}{Shared environmental path coefficient}

\item{e}{Unique environmental path coefficient}

\item{ct}{Cultural Transmission - Parent genotype to child phenotype}

\item{si}{Sibling Interaction - Sibling 1 genotype to sibling 2 phenotype}

\item{x}{Sibling Interaction at the phenotypic level}

\item{nloci}{Number of diallelic loci}

\item{npgsloci}{Number of loci comprising the PGS}

\item{assortm}{Assortative mating - genetic correlation between the parents}
}
\value{
List with data frame of power estimates and data frame of parameter estimates
}
\description{
Function to run OpenMx models estimating cultural transmission and sibling interaction on simulated data
}
\examples{
gnome_mx_simulation(ct = .01, si = .025, npgsloci = 10)
}
