% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnome_power.R
\name{gnome_power}
\alias{gnome_power}
\title{Function to calculate power from the non-centrality parameter}
\usage{
gnome_power(alpha = 0.05, df, ncp)
}
\arguments{
\item{alpha}{Type II error}

\item{df}{Degrees of freedom}

\item{ncp}{Non-centrality parameter}
}
\value{
Power
}
\description{
Function to calculate power from the non-centrality parameter
}
\examples{
gnome_power(df = 1, ncp = 10)
gnome_power(alpha = .01, df = 1, ncp = 5)
}
