% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powchi.R
\name{powchi}
\alias{powchi}
\title{Function to calculate power from the non-centrality parameter}
\usage{
powchi(alpha = 0.05, df, ncp)
}
\arguments{
\item{alpha}{Type II error}

\item{df}{Degrees of freedom}

\item{ncp}{Non-centrality parameter}
}
\value{
Power
}
\description{
Function to calculate power from the non-centrality parameter
}
\examples{
powchi(df = 1, ncp = 10)
powchi(alpha = .01, df = 1, ncp = 5)
}
