% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnome_effect.R
\name{gnome_effect}
\alias{gnome_effect}
\title{Function to calculate variance increase due to AC covariance}
\usage{
gnome_effect(a, c, e, g, b)
}
\arguments{
\item{a}{Additive genetic variance}

\item{c}{Shared environment variance}

\item{e}{Non-shared environmental variance}

\item{g}{Variance due to cultural transmission}

\item{b}{Variance due to sibling interaction}
}
\value{
Proportion of variance increase
}
\description{
Function to calculate variance increase due to AC covariance
}
\examples{
gnome_effect(a = sqrt(.4), c = sqrt(.3), e = sqrt(.3), g = 0, b = sqrt(.05))
gnome_effect(a = sqrt(.4), c = sqrt(.3), e = sqrt(.3), g = sqrt(.05), b = 0)
gnome_effect(a = sqrt(.4), c = sqrt(.3), e = sqrt(.3), g = sqrt(.05), b = sqrt(.05))
}
