% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnome_effect.R
\name{gnome_effect}
\alias{gnome_effect}
\title{Function to calculate variance increase due to AC covarianceuse}
\usage{
gnome_effect(a, c, e, g, b)
}
\arguments{
\item{a}{Additive genetic variance}

\item{c}{Shared environment variance}

\item{e}{Non-shared environmental variance}

\item{g}{Variance due to cultural transmission}

\item{b}{Variance due to sibling interaction}
}
\value{
Proportion of variance increase
}
\description{
Function to calculate variance increase due to AC covarianceuse
}
\examples{
gnome_effect(a = .4, c = .3, e = .3, g = 0, b = .05)
gnome_effect(a = .4, c = .3, e = .3, g = .05 , b = 0)
gnome_effect(a = .4, c = .3, e = .3, g = .05, b = .05)
}
